using an npm package to provide service call functions to client
