+++ .gitignore
@@ -8,6 +8,5 @@ public
 # Dependencies
 node_modules
 
-#Server files (Private Key/Certificate for HTTPS setup on backend)
-*.pem
-backend/serverFiles
+#Security files (Private Key,Certificate, etc for backend)
+backend/security
