+++ References.txt
@@ -7,6 +7,7 @@
 -firebase documentation
 -nodejs documentation
 -MDN js documentation
+-npm documentation
 
 
 Stack Overflow posts:
