+++ backend/client-pkg/index.js
@@ -1,0 +1,5 @@
+const HelloService = require('./services/HelloService');
+
+module.exports = {
+  HelloService,
+};
