+++ backend/client-pkg/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "comake-services",
+  "version": "1.0.1",
+  "description": "Client-side service code for CoMake",
+  "main": "index.js",
+  "scripts": {
+    "start": "node index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/calebhsu/CoMake.git"
+  },
+  "author": "CoMake Team",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/calebhsu/CoMake/issues"
+  },
+  "homepage": "https://github.com/calebhsu/CoMake#readme",
+  "devDependencies": {
+    "eslint": "^3.10.1",
+    "eslint-config-airbnb": "^13.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^2.2.3",
+    "eslint-plugin-react": "^6.7.1"
+  },
+  "dependencies": {
+    "http": "0.0.0"
+  }
+}
