+++ website/src/ServiceEndpoints.jsx
@@ -1,0 +1,8 @@
+const HELLO_SERVICE = {
+  host: 'localhost',
+  port: 8443,
+};
+
+export default {
+  HELLO_SERVICE,
+};
