--- website/src/services/Endpoints.jsx
@@ -1,10 +1,0 @@
-export const HELLO_SERVICE = {
-  host: 'localhost',
-  port: 8443,
-};
-
-export const Endpoints = {
-  HELLO_SERVICE,
-};
-
-export default Endpoints;
